dapp/requirements.txt
astroid==2.15.3
autopep8==2.0.2
blinker==1.6.2
click==8.1.3
codespell==2.2.4
dill==0.3.6
Flask==2.1.2
Flask-Classful==0.14.2
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==3.0.1
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
p2pnetwork==1.2
platformdirs==3.2.0
pony==0.7.16
pycodestyle==2.10.0
pycryptodome==3.17
pylint==2.17.2
tomli==2.0.1
tomlkit==0.11.7
typing_extensions==4.5.0
Werkzeug==2.1.2
wrapt==1.15.0
